T#777: Extract forge (Routine Tracker) to src/forge/routes.ts
Phase 1.12 of Backend Modularization (Library #102) — FINAL Phase 1 phase.
Mechanical move, no logic changes.

**Moved to src/forge/routes.ts (~1714L)**:
- CREATE TABLE: routine_logs, exercises, personal_records, photos
- Schema migrations (logged_at normalization, link_id column)
- Helpers: detectImageType, validateWorkoutData, parseExerciseName, parseExerciseString
- 25+ forge routes:
  - GET /api/routine/logs, today, weight, blood-pressure, exercise-summary, prs, workout-trends, body-composition, stats, summary
  - GET/POST /api/routine/exercises + seed
  - GET/POST /api/routine/personal-records + seed
  - GET /api/routine/photos
  - POST /api/routine/logs (workout/meal/measurement intake)
  - PATCH /api/routine/logs/:id, DELETE /api/routine/logs/:id, GET /api/routine/logs/deleted, PATCH /api/routine/logs/:id/restore
  - POST /api/routine/photo/upload, GET /api/routine/photo/:filename
  - POST /api/routine/import/{alpha-progression,alpha-measurements,hevy/sync}
  - POST /api/webhooks/hevy

**Cross-domain helper retained in server.ts**:
- isForgeAuthorized + FORGE_BEAST_MODES kept inline (~45L) — used by
  integrations module's /api/withings/devices gate via helpers DI.
  Forge routes themselves use the closure-local copy from the verbatim
  forge_body. T#788 cleanup pass may dedupe.

**Exports**:
- registerForgeRoutes(app, sqlite, { hasSessionAuth, isTrustedRequest, wsBroadcast })

**Imports**:
- timingSafeEqual from crypto (Hevy webhook signature verify)
- ORACLE_DATA_DIR from ../config.ts (photo upload paths)
- detectImageType inlined as module-level (de-duplicated from server.ts)

**Server.ts changes**:
- Added import after specs import
- Removed L6288-7963 forge block (~1676 lines) but RETAINED L6360-6404 helper inline
- Net: server.ts down ~1669 lines (was 8,329, now 6,660)

**Verification**:
- Boot test: server starts clean. All forge routes registered via DI.
- tsc --noEmit: zero new errors
- No logic changes: all auth gates preserved (isForgeAuthorized read/write modes,
  Hevy webhook signature verify, photo upload size limits)

**🎉 PHASE 1 COMPLETE — server.ts -5,303 lines / 44% reduction in 7 phases over ~2h45m**:
- PR #76 T#771 search: -455
- PR #77 T#772 scheduler: -660
- PR #78 T#773 daemons: -378
- PR #79 T#774 board: -555
- PR #80 T#775 integrations: -962
- PR #81 T#776 specs: -637
- PR #82 T#777 forge: -1669
- server.ts: 11,963 → 6,660 (-5,303 lines, -44%)
